requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
beautifulsoup4
selenium
requests
pymongo

